\name{MultiRD.predict.prop}
\alias{MultiRD.predict.prop}
\title{predict cell type proportions via MultiRD}
\description{
To extract optimal cell type proportions via MultiRD algorithm
}
\usage{
MultiRD.predict.prop(MultiRD.output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MultiRD.output}{An object from MultiRD.onegroup}
}
\value{Estimated cell type proportions via MultiRD}
