\name{MuRD.predict.prop}
\alias{MuRD.predict.prop}
\title{predict cell type proportions via MuRD}
\description{
To extract optimal cell type proportions via MuRD algorithm
}
\usage{
MuRD.predict.prop(murd.output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{murd.output}{An object from MuRD.onegroup}
}
\value{Estimated cell type proportions via MuRD}
