\name{pop.ct.prop.subj}
\alias{pop.ct.prop.subj}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
calculate population-level cell type proportions from subject-level cell type proportions
}
\description{
calculate population-level cell type proportions from subject-level cell type proportions
}
\usage{
pop.ct.prop.subj(
subj.prop
)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{subj.prop}{subject-level cell type proportions}
}
\value{
population-level cell type proportions
}
