\name{reffree.predict.prop}
\alias{reffree.predict.prop}
\title{predict cell type proportions via reference-free approach}
\description{
To extract optimal cell type proportions via reference-free algorithm
}
\usage{
reffree.predict.prop(MultiRD.output)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MultiRD.output}{An object from reffree.onegroup}
}
\value{Estimated cell type proportions via reffree}
